refactor(ui-grid): extract useGridKeyboardSelection
